+++ Stark/Stark/Controllers/ReviewsController.cs
@@ -22,15 +22,16 @@ namespace Stark.Controllers
         // GET: Reviews
         public async Task<IActionResult> Index(string sortOrder,string searchString)
         {
-            ViewData["NameSortParm"] = String.IsNullOrEmpty(sortOrder) ? "plate_desc" : "";
-            ViewData["DateSortParm"] = sortOrder == "Date" ? "date_desc" : "Date";
-            ViewData["CurrentFilter"] = searchString;
+            ViewData["NameSortParm"] = String.IsNullOrEmpty(sortOrder) ? "plate_desc" : ""; //used for sort button in the View
+            ViewData["DateSortParm"] = sortOrder == "Date" ? "date_desc" : "Date";//used for sort button in the View
+            ViewData["CurrentFilter"] = searchString; //used for search button in the View
             IQueryable<Review> starkContext = _context.Review.Include(r => r.Badge).Include(r => r.Licence);
-            if (!String.IsNullOrEmpty(searchString))
+            if (!String.IsNullOrEmpty(searchString)) //if anything was entered in the search box
             {
-                starkContext =_context.Review.Include(r => r.Badge).Include(r => r.Licence).Where(r => r.Licence.Plate.Contains(searchString.Trim())||r.Badge.Title.Contains(searchString.Trim())||r.UserIp.Contains(searchString.Trim())||((BadgeType)r.Badge.Type).ToString().Contains(searchString.Trim()));
+                //context passed to view is now limited to only the records that match searchString - operator OR used so all the columns can be filtered
+                starkContext =_context.Review.Include(r => r.Badge).Include(r => r.Licence).Where(r => r.Licence.Plate.ToUpper().Contains(searchString.ToUpper().Trim())||r.Badge.Title.ToUpper().Contains(searchString.ToUpper().Trim())||r.UserIp.Contains(searchString.Trim())||((BadgeType)r.Badge.Type).ToString().Contains(searchString.Trim())||r.CreateDate.Date.ToString().Contains(searchString.Trim()));
             }
-            switch (sortOrder)
+            switch (sortOrder) //used to determine which way the context is Ordered 
             {
                 case "plate_desc":
                     starkContext = starkContext.OrderByDescending(s => s.Licence.Plate);
@@ -56,7 +57,6 @@ namespace Stark.Controllers
             {
                 return NotFound();
             }
-
             var review = await _context.Review
                 .Include(r => r.Badge)
                 .Include(r => r.Licence)
@@ -68,8 +68,36 @@ namespace Stark.Controllers
 
             return View(review);
         }
+        // GET: Reviews/Create when the Plate already exists in the DB
+        public IActionResult CreateExists(string sequence, bool? saveChangesError = false)
+        {
+            //"badges" used for the View - all the field of a Badge record are now visible in the dropdown
+            var badges = _context.Badge.Select(m => new { Text = m.Title + " -- " + (BadgeType)Enum.ToObject(typeof(BadgeType), m.Type) + " -- " + m.Description, Value = m.BadgeId }).ToList(); //for the View - all the field of a Badge record are visible
 
-        // GET: Reviews/Create
+            //"aux" - selects the plate from the db (matching it with the parameter "sequence" that came from Cars POST redirect
+            var aux = _context.Cars.Where(r => r.Plate == sequence);
+            ViewData["BadgeId"] = new SelectList(badges, "Value", "Text");
+            ViewData["LicenceId"] = new SelectList(aux, "LicenceId", "Plate");
+            if (saveChangesError.GetValueOrDefault()) //spam check
+            {
+                ViewData["ErrorMessage"] ="You've recently reviewed this driver. You can review the same driver once every 72 hours.";
+            }
+
+            return View();
+        }
+
+        // GET: Reviews/Create when the Plate did not exist in the DB before the Upload POST call (that redirected to a Cars Create POST call)
+        public IActionResult CreateNew(int id)
+        {
+            //"badges" used for the View - all the field of a Badge record are now visible in the dropdown
+            var badges = _context.Badge.Select(m => new { Text = m.Title + " -- " + (BadgeType)Enum.ToObject(typeof(BadgeType), m.Type) + " -- " + m.Description, Value = m.BadgeId }).ToList();
+            //"aux" - selects the (freshly added) plate from the db (matching it with the parameter "id" that came from Cars POST redirect
+            var aux = _context.Cars.Where(r => r.LicenceId == id);
+            ViewData["BadgeId"] = new SelectList(badges, "Value", "Text");
+            ViewData["LicenceId"] = new SelectList(aux, "LicenceId", "Plate");
+            return View();
+        }
+        // GET: Reviews/Create - only called when adding Reviews "the admin way"
         public IActionResult Create()
         {
             var badges = _context.Badge.Select(m => new { Text = m.Title + " -- " + (BadgeType)Enum.ToObject(typeof(BadgeType), m.Type) + " -- " + m.Description, Value = m.BadgeId }).ToList();
@@ -85,15 +113,22 @@ namespace Stark.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("ReviewId,LicenceId,BadgeId,CreateDate,UserIp")] Review review)
         {
+            //"badges" used for the View - all the field of a Badge record are now visible in the dropdown
             var badges = _context.Badge.Select(m => new { Text = m.Title + " -- " + (BadgeType)Enum.ToObject(typeof(BadgeType), m.Type) + " -- " + m.Description, Value = m.BadgeId }).ToList();
+            //userip  - local variable used to get the IP of the user to be then assigned to the UserIp property of Review
             string userip = System.Net.Dns.GetHostEntry(System.Net.Dns.GetHostName()).AddressList.GetValue(1).ToString();
             review.UserIp = userip;
-            review.CreateDate = DateTime.Now;
-            if (ModelState.IsValid)
+            //spam - used to determine if the same user (same ip) has already reviewed the same Car (same LicenseId) in the last 72 hours
+            bool spam = _context.Review.Any(s => s.LicenceId == review.LicenceId && s.UserIp == review.UserIp && (review.CreateDate.Subtract(s.CreateDate).TotalDays<3.0d));
+            if ((ModelState.IsValid)&&(!spam))
             {
                 _context.Add(review);
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
+            }
+            if (spam==true)
+            {
+                return RedirectToAction("CreateExists", new { saveChangesError = true }); //for generating a spam error message in the View
             }
             ViewData["BadgeId"] = new SelectList(badges, "Value", "Text");
             ViewData["LicenceId"] = new SelectList(_context.Cars, "LicenceId", "Plate", review.LicenceId);
@@ -157,7 +192,7 @@ namespace Stark.Controllers
         }
 
         // GET: Reviews/Delete/5
-        public async Task<IActionResult> Delete(int? id)
+        public async Task<IActionResult> Delete(int? id, bool? saveChangesError = false)
         {
             if (id == null)
             {
@@ -172,6 +207,11 @@ namespace Stark.Controllers
             {
                 return NotFound();
             }
+            if (saveChangesError.GetValueOrDefault())
+            {
+                ViewData["ErrorMessage"] =
+                    "Delete failed";
+            }
 
             return View(review);
         }
@@ -181,10 +221,21 @@ namespace Stark.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var review = await _context.Review.FindAsync(id);
-            _context.Review.Remove(review);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            var review = await _context.Review.AsNoTracking().SingleOrDefaultAsync(m =>m.ReviewId ==id);
+            if (review==null)
+            {
+                return RedirectToAction(nameof(Index));
+            }
+            try
+            {
+                _context.Review.Remove(review);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            catch (DbUpdateException)
+            {
+                return RedirectToAction(nameof(Delete), new { id = id, saveChangesError = true });
+            }
         }
 
         private bool ReviewExists(int id)
